Added ActionBar features The Control, Connection List, Connection Edit, and File Explorer activites are now blessed to have ActionBar icons! On devices that don't have an ActionBar, they will still appear in the Context Menu. Imported stock icons to use in the main control activity. Reworked the Connection List activity for more intuitive management of the connections, namely that there are now ActionBar actions that impact the currently selected ("used") connection. The existing method of Tap-and-Hold still works just fine.
Fixed display issues for the Textfield section on devices prior to Honeycomb (It wouldn't move below the buttons, but stay overlayed).

Server:
Fixed an issue where Windows wouldn't attempt to launch a file with Spaces in the Filename.
Made the error messages displayed when attempting to launch a file more friendly.